% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMRT.R
\name{calcGT}
\alias{calcGT}
\title{calcGT}
\usage{
calcGT(Ta, Td, WV, S = 336, Fdiff, Fd, Z, P)
}
\arguments{
\item{Ta}{numeric, Air temperature (°C)}

\item{Td}{numeric, dew point temperature (°C)}

\item{WV}{numeric, wind speed (meters per second)}

\item{S}{numeric, solar radiation (Watts per meter squared)}

\item{Fdiff}{numeric, diffuse solar radiation (Watts per meter squared)}

\item{Fd}{numeric, direct solar radiation (Watts per meter squared)}

\item{Z}{numeric, zenith angle of sun (degrees)}

\item{P}{numeric, pressure in millibars}
}
\value{
globe temperature (°C)
}
\description{
calculates Globe Temperature (Tg)
}
\examples{
# From APPENDIX of Dimiceli, V.E.; Piltz, S.F.; Amburn, S.A.
# Estimation of Black Globe Temperature for Calculation of the
# Wet Bulb Globe Temperature Index.
# In Proceedings of the Proceedings of the World Congress on Engineering
# and Computer Science; 2011; Vol. II.
gt1 <- calcGT(Ta=30, Td=20.556, WV=2.681666667, S=336, Fdiff=0.3333, Fd=0.6666, Z=38.44, P=992.83 )
gt2 <- calcGT(Ta=33.889,  Td=24.444, WV=11263/3600, S=754, Fdiff=0.25, Fd=0.75, Z=36.65, P=981.94 )
}
