% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMRT.R
\name{calcMRT}
\alias{calcMRT}
\title{calcMRT}
\usage{
calcMRT(Tg = 21, WV = 0, Ta = 21)
}
\arguments{
\item{Tg}{numeric, globe temperature in °C, DEFAULT=21}

\item{WV}{numeric, air velocity at the level of the globe in meters
per second , DEFAULT=0}

\item{Ta}{numeric, Air temperature in °C, DEFAULT=21}
}
\value{
mean radiant temperature in °C
}
\description{
calculates mean radiant temperature
}
\examples{
Ta=30; WV=2.681666667;
#Tg <- calcGT(Ta=Ta, Td=20.556, WV=WV, S=336, Fdiff=0.3333, Fd=0.6666, Z=38.44, P=992.83 )
#calcMRT(Tg=21, WV=0, Ta=21 )
}
