% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMRT.R
\name{calcMRT2}
\alias{calcMRT2}
\title{calcMRT2}
\usage{
calcMRT2(Ta, Td, WV, S, Fdiff, Fd, Z, P)
}
\arguments{
\item{Ta}{numeric, Air temperature (°C)}

\item{Td}{numeric, dew point temperature (°C)}

\item{WV}{numeric, wind speed (meters per second)}

\item{S}{numeric, solar radiation (Watts per meter squared)}

\item{Fdiff}{numeric, diffuse solar radiation (Watts per meter squared)}

\item{Fd}{numeric, direct solar radiation (Watts per meter squared)}

\item{Z}{numeric, zenith angle of sun (degrees)}

\item{P}{numeric, pressure in millibars
per second , DEFAULT=0}
}
\value{
mean radiant temperature in °C
}
\description{
calculates mean radiant temperature estimating globe temperature
}
\examples{
Ta=30; WV=2.681666667;

}
