% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclim.R
\docType{data}
\name{dataclim}
\alias{dataclim}
\title{PET examples data}
\format{
\subsection{\code{dataclim}}{

A data frame with 5 rows and 6 columns:
\describe{
\item{scenario}{Country name}
\item{ta,tmrt}{Air Temperature and Mean Radiant Temperature in °C}
\item{v}{Wind velocity in m/s}
\item{VP}{Vapour Pressure in hPa }
\item{PET}{ physiological equivalent temperature (PET) in  equiv. °C }
}
}
}
\source{
\url{https://doi.org/10.1007/s004840050118}
}
\usage{
dataclim
}
\description{
Examples from Höppe, P. The physiological equivalent temperature – a universal index
for the biometeorological assessment of the thermal environment.
Int J Biometeorol 43, 71–75 (1999). zzz- \url{https://doi.org/10.1007/s004840050118}
}
\keyword{datasets}
