% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{declination}
\alias{declination}
\title{declination}
\usage{
declination(jd)
}
\arguments{
\item{jd}{Julian Day.}
}
\value{
Declination in degrees and decimal fraction
}
\description{
Computes the declination of the Sun for a given Julian Day.
}
\examples{
 declination(JDymd(2019,1,1))
}
\author{
Javier G. Corripio
}
