% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{degrees}
\alias{degrees}
\title{degrees}
\usage{
degrees(radian)
}
\arguments{
\item{radian}{angle in radians}
}
\value{
angle in degrees
}
\description{
Converts angles from radians to degrees
}
\examples{
 degrees(seq(0,2*pi,pi/2))
}
\author{
Javier G. Corripio
}
