% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{eqtime}
\alias{eqtime}
\title{eqtime}
\usage{
eqtime(jd)
}
\arguments{
\item{jd}{numeric Julian Day}
}
\value{
Equation of time in minutes.
}
\description{
Computes the equation of time for a given Julian Day.
}
\examples{
 # plot the equation of time for 2013 at daily intervals
jdays = seq(ISOdate(2013,1,1),ISOdate(2013,12,31),by='day')
jd = JD(jdays)
plot(eqtime(jd))
abline(h=0,col=8)

}
\author{
Javier G. Corripio
}
