% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{hourangle}
\alias{hourangle}
\title{hourangle}
\usage{
hourangle(jd, longitude, timezone)
}
\arguments{
\item{jd}{Julian Day.}

\item{longitude}{Longitude.}

\item{timezone}{Timezone in hours, west of Greenwich is negative.}
}
\value{
Hour angle
}
\description{
Hour angle, internal function for solar position.
}
\author{
Javier G. Corripio
}
