% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{JD}
\alias{JD}
\title{JD}
\usage{
JD(x, inverse=FALSE)
}
\arguments{
\item{x}{POSIXct object.}

\item{inverse}{Logical. If false (default) returns the Julian Days
corresponding to given dates. If TRUE returns the date corresponding
to input Julian days}
}
\value{
Julian Day
}
\description{
Computes Julian Day from dates as POSIXct object.
}
\details{
Class "POSIXct" represents the (signed) number of seconds since the
beginning of 1970 (in the UTC timezone) as a numeric vector, and Julian Day is
the number of days since January 1, 4713 BCE at noon UTC, so the Julian Day
is calculated as numeric(POSIXct)+2440587.5 days.
}
\examples{
 JD(Sys.time())
}
\author{
Javier G. Corripio
}
