% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{JDymd}
\alias{JDymd}
\title{JDymd}
\usage{
JDymd(year,month,day,hour=12,minute=0,sec=0)
}
\arguments{
\item{year}{numeric year}

\item{month}{1-12: number of the month.}

\item{day}{1-31: day of the month.}

\item{hour}{0-23: hour of the day.}

\item{minute}{0-59: minutes.}

\item{sec}{0-59: seconds.}
}
\value{
Julian Day, or number of days since January 1, 4713 BCE at noon UTC
}
\description{
Computes Julian Day from a given date.
}
\examples{
 declination(JDymd(2019,1,1))
}
\author{
Javier G. Corripio
}
