% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2raster.R
\name{matrix2raster}
\alias{matrix2raster}
\title{matrix2raster}
\usage{
matrix2raster(matrix, raster = NULL)
}
\arguments{
\item{matrix}{2d matrix with cell values}

\item{raster}{Default \code{NULL}. If a raster is provided, CRS and}
}
\value{
raster object from {terra} library
}
\description{
converts a matrix to a terra::rast raster valuesì
}
\examples{
#
}
