% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMRT.R
\name{mrt}
\alias{mrt}
\title{calcMRT}
\usage{
mrt(air_temperature, sunaltitude, solar_radiation_wm2, Fd = 1)
}
\arguments{
\item{air_temperature}{numeric, Air temperature in °C, DEFAULT=21}

\item{sunaltitude}{numeric, sun altitude from the horizong (90-zenith),
DEFAULT=50 - used to determine the surface projection factor for the
standing human body with rotational symmetry (i.e. independent of
azimuth of the sun)}

\item{solar_radiation_wm2}{Solar radiation in W/m2 - at 46° latitude it
ranges from 0 to 1000 W/m2}

\item{Fd}{the fraction of direct solar radiation reaching the body,
from 0 (body is not illuminated by the sun)
to 1 (fully explosed to solar radiation)}
}
\value{
mean radiant temperature in °C
}
\description{
calculates mean radiant temperature from air temperature
and
}
\examples{
mrt(air_temperature=21, sunaltitude=50, solar_radiation_wm2=100  )
}
