% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETcorrected.R
\name{PETcorrected}
\alias{PETcorrected}
\title{Calculate PET}
\usage{
PETcorrected(Tair, Tmrt, v_air, pvap, M_activity, icl)
}
\arguments{
\item{Tair}{numeric DEFAULT 21}

\item{Tmrt}{numeric DEFAULT 21}

\item{v_air}{numeric DEFAULT 0.1}

\item{pvap}{numeric DEFAULT 21}

\item{M_activity}{numeric DEFAULT 80}

\item{icl}{numeric DEFAULT 0.9}
}
\value{
PET value
}
\description{
Calculate PET
}
\examples{
PETcorrected(1,2)
}
