% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare Example Data}
\usage{
prepareData(force = FALSE)
}
\arguments{
\item{force, }{boolean, DEFAULT = TRUE.}
}
\value{
boolean with TRUE on success
}
\description{
Prepares data by downloading and creating an environmental variable that you
can access by e<-environment(rPET::prepareData) and e$dataenv
}
\examples{
#First we ray trace the Monterey Bay dataset.
}
