% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{radians}
\alias{radians}
\title{radians}
\usage{
radians(degree)
}
\arguments{
\item{degree}{angle in degrees}
}
\value{
angle in radians.
}
\description{
Converts angles to radians from degrees
}
\examples{
 radians(90)
}
\author{
Javier G. Corripio
}
