% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPET-package.R
\docType{package}
\name{rPET-package}
\alias{rPET}
\alias{rPET-package}
\title{rPET: Mapping Physiological Equivalent Temperature}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fpirotti/rPET}
}

}
\author{
\strong{Maintainer}: Francesco Pirotti \email{francesco.pirotti@unipd.it} (\href{https://orcid.org/0000-0002-4796-6406}{ORCID})

}
\keyword{internal}
