% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{sunpos}
\alias{sunpos}
\title{sunpos}
\usage{
sunpos(sunv)
}
\arguments{
\item{sunv}{coordinates x, y, z of the unit vector in the direction
of the sun.}
}
\value{
A matrix of azimuth and zenith angles in degrees.
}
\description{
Returns a matrix of azimuth and zenith angles of the sun given the
unit vectors from the observer to the direction of the sun.
}
\examples{
 sunpos(matrix( (1:6)/6, nrow=2, ncol=3))
}
\author{
Javier G. Corripio
}
