% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunpos.R
\name{sunvector}
\alias{sunvector}
\title{sunvector}
\usage{
sunvector(jd, latitude, longitude, timezone)
}
\arguments{
\item{jd}{Julian Day and decimal fraction.}

\item{latitude}{Latitude of observer in degrees and decimal fraction.}

\item{longitude}{Longitude of observer in degrees and decimal fraction.}

\item{timezone}{Time zone, west is negative.}
}
\value{
3 column matrix with the x, y , z coordinates of the sun vector.
}
\description{
Calculates a unit vector in the direction of the sun from
the observer position.
}
\details{
To calculate the sunvector to the nearest hour, give the Julian Day
with a precission better than 1/24; to approximate it to the nearest minute
use decimal fractions smaller than 1/(24*60), and so on.
}
\examples{
 sunvector(JD(Sys.time()),51.4778,-0.0017,0)
}
\author{
Javier G. Corripio
}
