% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETcorrected.R
\name{PETcorrected}
\alias{PETcorrected}
\title{Calculate PET}
\usage{
PETcorrected(
  Tair = 21,
  Tmrt = 21,
  v_air = 0.1,
  pvap = 21,
  M_activity = 80,
  icl = 0.9
)
}
\arguments{
\item{Tair}{numeric, air temperature in degrees celsius DEFAULT 21}

\item{Tmrt}{numeric DEFAULT 21}

\item{v_air}{numeric, air speed meters per second DEFAULT 0.1}

\item{pvap}{numeric, vapour pressure as relative umidity (\%) DEFAULT 21}

\item{M_activity}{numeric DEFAULT 80}

\item{icl}{numeric Clothing level DEFAULT 0.9}
}
\value{
PET value
}
\description{
Calculate PET
}
\examples{
PETcorrected(21,21,0.1,21,80,0.9)
}
