% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETcorrected.R
\name{PETcorrected}
\alias{PETcorrected}
\title{Calculate PET}
\usage{
PETcorrected(
  Tair = 21,
  Tmrt = 21,
  v_air = 0.1,
  pvap = 21,
  M_activity = 80,
  icl = 0.9
)
}
\arguments{
\item{Tair}{numeric, air temperature in degrees celsius DEFAULT=21}

\item{Tmrt}{numeric, Mean radiant temperature in degrees celsius DEFAULT=21}

\item{v_air}{numeric, air speed meters per second DEFAULT=0.1}

\item{pvap}{numeric, vapour pressure as relative umidity (\\%) DEFAULT=21}

\item{M_activity}{numeric, metabolic rate (W/m²), DEFAULT=80 (about 1.4 met) see\preformatted{ https://en.wikipedia.org/wiki/Thermal_comfort#Metabolic_rate

 Below some values (met) - (1 met = 58.2 W/m² )
}

Sleeping: 0.7

Reclining: 0.8

Seated, quiet, Reading, seated, Writing: 1.0 (58.2 W/m²)

Typing: 1.1 (64 W/m²)

Standing, relaxed,  seated: 1.2 (70 W/m²)

Walking slow: 1.4  (81 W/m²)

Driving a car: 1.5  (87 W/m²)

Walking medium speed: 1.7  (99 W/m²)

Walking 2mph (3.2km/h): 2.0  (116 W/m²)

Light machine work: 2.2  (128 W/m²)

Walking 3mph (4.8km/h): 2.6

House cleaning: 2.7

Driving, heavy vehicle: 3.2

Dancing: 3.4

Walking 4mph (6.4km/h): 3.8

Heavy machine work: 4.0}

\item{icl}{numeric, Clothing level DEFAULT 0.9}
}
\value{
PET value
}
\description{
Calculate PET, Physiological Equivalent Temperature, check also https://comfort.cbe.berkeley.edu/
}
\examples{
PETcorrected(Tair=21, Tmrt=21, v_air=0.1, pvap=50, M_activity=80, icl=0.9)
}
