% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETcorrected.R
\name{PETcorrected}
\alias{PETcorrected}
\title{Calculate PET}
\usage{
PETcorrected(
  Tair = 21,
  Tmrt = 21,
  v_air = 0.1,
  rh = 50,
  M_activity = 80,
  icl = 0.9,
  bodyPosition = "standing",
  sex = "male",
  age = 35,
  mbody = 75,
  ht = 1.8,
  verbose = FALSE
)
}
\arguments{
\item{Tair}{numeric, air temperature in degrees celsius, DEFAULT=21}

\item{Tmrt}{numeric, Mean radiant temperature in degrees celsius, DEFAULT=21}

\item{v_air}{numeric, air speed meters per second, DEFAULT=0.1}

\item{rh}{numeric, vapour pressure as relative humidity in percentage, will be converted to hPa using the Clausius-Clapeyron equation, DEFAULT=50}

\item{M_activity}{numeric, metabolic rate (W/m²), DEFAULT=80 (about 1.4 met) see table below.}

\item{icl}{numeric, clothing level in clot values, DEFAULT=0.9, 0=naked, 2=fully clothed with jackets and hats}

\item{bodyPosition}{character, can be "sitting", "crouching" or "standing", DEFAULT=standing}

\item{sex}{character, male/female, DEFAULT=male}

\item{age}{numeric, age of the person, DEFAULT=35}

\item{mbody}{numeric, body weight in kg, DEFAULT=75}

\item{ht}{numeric, person's height in meters DEFAULT=1.75}

\item{verbose}{boolean, extra messages
\url{https://en.wikipedia.org/wiki/Thermal_comfort#Metabolic_rate}

Below some values (met) - (1 met = 58.2 W/m² )
\itemize{
\item Sleeping: 0.7 met,  (41 W/m²)
\item Reclining: 0.8 met, (47 W/m²)
\item Seated, quiet, Reading, seated, Writing: 1.0 (58.2 W/m²)
\item Typing: 1.1 met, (64 W/m²)
\item Standing, relaxed,  seated: 1.2 (70 W/m²)
\item Walking slow: 1.4 met,  (81 W/m²)
\item Driving a car: 1.5 met,  (87 W/m²)
\item Walking medium speed: 1.7 met,  (99 W/m²)
\item Walking 2mph (3.2km/h): 2.0 met,  (116 W/m²)
\item Light machine work: 2.2 met,  (128 W/m²)
\item Walking 3mph (4.8km/h): 2.6 met,  (151 W/m²)
\item House cleaning: 2.7 met,  (157 W/m²)
\item Driving, heavy vehicle: 3.2 met,  (186 W/m²)
\item Dancing: 3.4 met,  (198 W/m²)
\item Walking 4mph (6.4km/h): 3.8 met,  (221 W/m²)
\item Heavy machine work: 4.0 met,  (233 W/m²)
}}
}
\value{
PET value
}
\description{
Calculate PET, Physiological Equivalent Temperature, check also https://comfort.cbe.berkeley.edu/
}
\examples{
# Dataset from
PETc <- PETcorrected(rPET::dataclim$Ta, rPET::dataclim$Tmrt,
                     rPET::dataclim$v, rPET::dataclim$rh_perc)

# It can be applied vectorized, below and example with three air
# temperatures and wind speeds:
PETcorrected( Tair = 20, Tmrt=21, v_air=0, rh=20 )
PETcorrected( Tair = 25, Tmrt=21, v_air=0.1, rh=20 )
PETcorrected( Tair = 30, Tmrt=21, v_air=0.5, rh=20 )
# a faster way:
PETcorrected( Tair = c(20,25,30), Tmrt=21, v_air=c(0, 0.1, 0.5), rh=20 )
}
